+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [unreleased]
 
+### Added
+
+ - `ident.h`:
+   - `ML99_charLit`, `ML99_CHAR_LIT` to convert a Metalang99 character to a C character literal.
+
 ## [1.7.0] - 2021-08-13
 
 ### Changed